+++ app/Http/Livewire/Admin/Role/Index.php
@@ -2,14 +2,16 @@
 
 namespace App\Http\Livewire\Admin\Role;
 
-
 use Illuminate\Support\Facades\Gate;
 use Livewire\Component;
 use Livewire\WithPagination;
-use App\Repositories\Contract\{ILog,IRole};
+use App\Repositories\Contract\{
+    ILog,
+    IRole
+};
 
-class Index extends Component
-{
+class Index extends Component {
+
     use WithPagination;
 
     public $category;
@@ -23,65 +25,76 @@ class Index extends Component
     public $searchTerm = null;
     public $sortColumnName = 'created_at';
     public $sortDirection = 'desc';
-    public $SelectPage=false;
-    public $typePage          ='roles';
+    public $SelectPage = false;
+    public $typePage = 'roles';
 
+    public function mount() {
+        if (!Gate::allows('show_role')) {
+            abort(403);
+        }
+    }
 
     public function createLog($data) {
-        
-        return  app()->make(ILog::class)->create($data);
+
+        return app()->make(ILog::class)->create($data);
     }
-    
-    
+
     public function getInterface() {
 
-        return  app()->make(IRole::class);
-
+        return app()->make(IRole::class);
     }
 
-    public function UpdatedSelectPage($value)  {  
-        $param =['url',$this->search,$this->count_data];
-        $this->mulitiSelect = $this->getInterface()->UpdatedSelectPage($value,$param);
-
+    public function UpdatedSelectPage($value) {
+        $param = ['url', $this->search, $this->count_data];
+        $this->mulitiSelect = $this->getInterface()->UpdatedSelectPage($value, $param);
     }
 
-    public function confirmRemoval($categoryId)  {
+    public function confirmRemoval($categoryId) {
         $this->IdBeingRemoved = $categoryId;
         $this->dispatchBrowserEvent('show-delete-modal');
-
     }
 
     public function confirmAllRemoval() {
         $this->dispatchBrowserEvent('show-form');
     }
 
-    public function deleteAll() {        
-        $this->getInterface()->deleteAll($this->mulitiSelect);
-        $this->createLog([
-           'user_id'     => auth()->user()->id, 
-           'actionType'  => 'Mass deletion', 
-           'url'         => $this->typePage, 
-        ]);
-        $this->mulitiSelect = [];
-        $this->SelectPage = false;
-        $this->dispatchBrowserEvent('hide-form');
-        $this->emit('toast', 'success', 'success !');
+    public function deleteAll() {
+        if (Gate::allows('edit_role')) {
+            $this->getInterface()->deleteAll($this->mulitiSelect);
+            $this->createLog([
+                'user_id' => auth()->user()->id,
+                'actionType' => 'Mass deletion',
+                'url' => $this->typePage,
+            ]);
+            $this->mulitiSelect = [];
+            $this->SelectPage = false;
+            $this->dispatchBrowserEvent('hide-form');
+            $this->emit('toast', 'success', 'success !');
+        } else {
 
+            $this->dispatchBrowserEvent('hide-form');
+            $this->emit('toast', 'warning', 'permission denied !');
+        }
     }
 
-    public function delete()  {
-        
-        $name = $this->getInterface()->find($this->IdBeingRemoved)->name;
-        
-        $this->getInterface()->delete($this->IdBeingRemoved);
-        
-        $this->createLog([
-           'user_id'     => auth()->user()->id, 
-           'actionType'  => 'delete '. $this->typePage, 
-           'url'         => $name, 
-        ]);
-        $this->dispatchBrowserEvent('hide-delete-modal');
-        $this->emit('toast', 'success', 'success !');
+    public function delete() {
+        if (Gate::allows('edit_role')) {
+            $name = $this->getInterface()->find($this->IdBeingRemoved)->name;
+
+            $this->getInterface()->delete($this->IdBeingRemoved);
+
+            $this->createLog([
+                'user_id' => auth()->user()->id,
+                'actionType' => 'delete ' . $this->typePage,
+                'url' => $name,
+            ]);
+            $this->dispatchBrowserEvent('hide-delete-modal');
+            $this->emit('toast', 'success', 'success !');
+        } else {
+
+            $this->dispatchBrowserEvent('hide-delete-modal');
+            $this->emit('toast', 'warning', 'permission denied !');
+        }
     }
 
     public function cancelDelete() {
@@ -112,13 +125,13 @@ class Index extends Component
 
         $this->readyToLoad = true;
     }
-    
-    public function render()
-    {
-        $data_info = $this->readyToLoad ? $this->getInterface()->all('label',$this->search)
-        ->orderBy($this->sortColumnName, $this->sortDirection)
-        ->latest()->paginate($this->count_data) : [];
+
+    public function render() {
+        $data_info = $this->readyToLoad ? $this->getInterface()->all('label', $this->search)
+                        ->orderBy($this->sortColumnName, $this->sortDirection)
+                        ->latest()->paginate($this->count_data) : [];
         $deleteItem = $this->mulitiSelect;
         return view('livewire.admin.role.index', compact('data_info', 'deleteItem'))->layout('layouts.admin');
     }
+
 }
